GPIO pin not showing in sensing menu.
In January 17 Raspbian release (and maybe earlier) the instructions for using the GPIO pin did not work for me. I had to configure the pin using "broadcast config2inpullup" to get the pin to show in the sensor value menu. Also, it shows up as "gpio2" rather than "pin3". Document updated accordingly.