Add pin config and correct drum
Changed "pin3" to "gpio2" to reflect previous commit changing the document text.
Document text states the drum as "48" but original graphic showed "61" so changed to "48".
This image is a bit rough around the edges. If needed I could make a cleaner one but this is good enough for my pupils to use.